Refactored package structure to work from command line.
* Top-level directories: postprocessors/, scripts/, and test/
* Within postprocessors, there is core/, exceptions/, utils/, and sandbox/
* Have yet to verify that this structure works with command line.
