+++ postprocessors/core/antennas_iq_to_rawacf.py
@@ -1,18 +1,19 @@
-# Copyright 2022 SuperDARN Canada, University of Saskatchewan
-
+# Copyright 2021 SuperDARN Canada, University of Saskatchewan
+# Author: Remington Rohel
 """
-This file contains functions for extracting sequence timestamps from any file type.
+This file contains functions for converting antennas_iq files
+to rawacf files.
 """
 from collections import OrderedDict
 from typing import Union
 
-from data_processing.convert_base import BaseConvert
+from postprocessors import BaseConvert, ProcessAntennasIQ2Bfiq, ProcessBfiq2Rawacf
 
 
-class ExtractTimestamps(BaseConvert):
+class ProcessAntennasIQ2Rawacf(BaseConvert):
     """
-    Class for extraction of sequence timestamps for bistatic experiments.
-    This class inherits from BaseConvert, which handles all functionality generic to postprocessing borealis files.
+    Class for conversion of Borealis antennas_iq files into rawacf files. This class inherits from
+    BaseConvert, which handles all functionality generic to postprocessing borealis files.
 
     See Also
     --------
@@ -20,12 +21,11 @@ class ExtractTimestamps(BaseConvert):
     BaseConvert
     ProcessAntennasIQ2Bfiq
     ProcessBfiq2Rawacf
-    ProcessAntennasIQ2Rawacf
 
     Attributes
     ----------
     infile: str
-        The filename of the input file.
+        The filename of the input antennas_iq file.
     outfile: str
         The file name of output file
     infile_structure: str
@@ -36,8 +36,8 @@ class ExtractTimestamps(BaseConvert):
         The desired structure of the output file. Same structures as above, plus 'dmap'.
     """
 
-    def __init__(self, infile: str, outfile: str, infile_type: str, outfile_type: str,
-                 infile_structure: str, outfile_structure: str):
+    def __init__(self, infile: str, outfile: str, infile_structure: str, outfile_structure: str,
+                 averaging_method: str = 'mean'):
         """
         Initialize the attributes of the class.
 
@@ -47,23 +47,22 @@ class ExtractTimestamps(BaseConvert):
             Path to input file.
         outfile: str
             Path to output file.
-        infile_type: str
-            Borealis filetype of input file. 'antennas_iq', 'bfiq', or 'rawacf'.
-        outfile_type: str
-            Borealis filetype of output file. 'antennas_iq', 'bfiq', or 'rawacf'.
         infile_structure: str
             Borealis structure of input file. Either 'array' or 'site'.
         outfile_structure: str
             Borealis structure of output file. Either 'array', 'site', or 'dmap'.
+        averaging_method: str
+            Method for averaging correlations across sequences. Either 'median' or 'mean'.
         """
-        super().__init__(infile, outfile, infile_type, outfile_type, infile_structure, outfile_structure)
+        super().__init__(infile, outfile, 'antennas_iq', 'rawacf', infile_structure, outfile_structure)
+        self.averaging_method = averaging_method
 
         self.process_file()
 
     @staticmethod
     def process_record(record: OrderedDict, averaging_method: Union[None, str], **kwargs) -> OrderedDict:
         """
-        Takes a record from a Borealis file and extract the sequence timestamps only.
+        Takes a record from an antennas_iq file process into a rawacf record.
 
         Parameters
         ----------
@@ -75,9 +74,9 @@ class ExtractTimestamps(BaseConvert):
         Returns
         -------
         record: OrderedDict
-            hdf5 record, containing only the sequence timestamps.
+            hdf5 record, with new fields required by rawacf data format
         """
-        new_record = OrderedDict()
-        new_record['sqn_timestamps'] = record['sqn_timestamps']
+        record = ProcessAntennasIQ2Bfiq.process_record(record, averaging_method)
+        record = ProcessBfiq2Rawacf.process_record(record, averaging_method)
 
-        return new_record
+        return record
