+++ postprocessors/sandbox/rawacf_lag0.py
@@ -8,7 +8,7 @@ from collections import OrderedDict
 from typing import Union
 import numpy as np
 
-from data_processing.convert_base import BaseConvert
+from postprocessors import BaseConvert
 
 
 class ReprocessRawacfLag0(BaseConvert):
